Allow to query non-existant sections
Sometimes it is convenient to query a library for a specific section and expect an empty result if the section does not exists. This was not possible because library_info ASSERTed the section adresses thus crashing the Application if a section was not found.
Fixed this by checking section begin and end and returning an emtpy vector of string if the section was not found.